requirements.txt
matplotlib
python-telegram-bot
schedule
